+++ packages/core/src/utils/index.js
@@ -2,7 +2,6 @@ export * from './utils'
 
 export { createColorStyles } from './createColorStyles'
 export { createTextStyles } from './createTextStyles'
-export { getSCMigrationRef } from './getSCMigrationRef'
 
 export { createTheme } from './createTheme'
 export { boxShadow } from './boxShadow'