+++ packages/core/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - pcln-design-system
 
-This log was last generated on Fri, 13 Nov 2020 20:04:31 GMT and should not be manually modified.
+This log was last generated on Tue, 17 Nov 2020 21:35:08 GMT and should not be manually modified.
+
+## 4.0.0
+Tue, 17 Nov 2020 21:35:08 GMT
+
+### Breaking changes
+
+- React floor bump to 16.10
+- use styled-components v4 as dev and peer dependency
+
+### Patches
+
+- Remove usages of mapProps and getSCMigrationRef
 
 ## 3.7.1
 Fri, 13 Nov 2020 20:04:31 GMT