+++ packages/modal/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log - pcln-modal
 
-This log was last generated on Fri, 23 Oct 2020 15:15:48 GMT and should not be manually modified.
+This log was last generated on Tue, 17 Nov 2020 21:35:08 GMT and should not be manually modified.
+
+## 4.0.0
+Tue, 17 Nov 2020 21:35:08 GMT
+
+### Breaking changes
+
+- React floor bump to 16.10
+- use styled-components v4 as dev and peer dependency
 
 ## 3.4.1
 Fri, 23 Oct 2020 15:15:48 GMT